+++ stevedore/tests/test_driver.py
@@ -10,6 +10,14 @@ def test_detect_plugins():
     assert names == ['t1']
 
 
+def test_call():
+    def invoke(ext, *args, **kwds):
+        return (ext.name, args, kwds)
+    em = driver.DriverManager('stevedore.test.extension', 't1')
+    result = em(invoke, 'a', b='C')
+    assert result == ('t1', ('a',), {'b': 'C'})
+
+
 def test_no_drivers():
     try:
         driver.DriverManager('stevedore.test.extension.none', 't1')