+++ docs/source/history.rst
@@ -6,6 +6,9 @@
 
   - Added dispatch managers for selecting among a set of plugins at
     runtime instead of load time.
+  - Added ``__call__`` method to
+    :class:`~stevedore.driver.DriverManager` so it can be invoked in a
+    more natural fashion for a single plugin.
 
 0.2